Improvements to direct-rendering CallbackMediaPlayer
Main change is optional automatic re-scaling of the video as the
container size changes. The re-scaling is accomplished by setting
the appropriate affine transform on the graphics context. This
scaling will also be reflected in any subsequent overlay rendering.

The rendering of the video is abstracted to a "painter" interface,
provided by default are a fixed-size painter and the aforementioned
scaling painter.

By default, the scaling version is used.

A client application can override these with sub-classes of these
painters, or their own bespoke implementation.

The above pertains only to the default behaviour provided by the
component - a client application is still free to provide their own
implementation of a RenderCallback to be in complete control of the
video rendering.