Further improvements to the CallbackMediaPlayerComponent
Now, when relying on the component to render the video, it is no
longer necessary to specify on creation the required size for the
video buffer and image used to render that buffer.

Instead, each time new video is played, when the buffer format is
invoked we use the supplied source video width and height to
create the exact-sized image and buffer that we need. This is then
set on the render callback adapter.

Remember, all scaling is done when the video is painted, so having
a buffer size that matches the source width and height makes sense.

It is also worth nothing that this implementation always uses a
fixed buffer format (RV32) which is 8-bit RGB and 8-bits of padding
for alpha - i.e. the format that matches that of a Java
BufferedImage.