Clean up the MediaFactory and MediaListFactory
Also add factory-type methods to Media and MediaList to create the
corresponding "ref" objects and vice versa.

When using the factory to create new instances, generally the
native media instance will be *duplicated* rather than *retained*.

This is so, for example, a client application could be managing its
media instances, and choose to play media with a set of media
options applied. These media options are applied to the duplicate
leaving the original media instance untouched.

If a client wants to create components that share the native media
instance, then they use the aforementioned factory-type methods
instead.