Improvements to the Renderer Discoverer component
Refactored component implementation to be consistent with how the
other high-level components have been implemented.

The component now maintains a list of discovered renderers which
can be queried (e.g. to display in a UI). The client application
should still rely on the event listeners to handle devices coming
and going.

Moved the actual setting of the renderer to the media player
rather than the renderer item.