Enable "persistent" media event listeners on media player
This means that a client application can add a MediaEventListener
to the media player, and have that event listener persist across
changes to the media.

The client is still free to add their own listeners to the media
if they want.

Corresponding changes made to the EmbeddedMediaPlayerComponent,
CallbackMediaPlayerComponent and AudioPlayerComponent.