Replace libvlc_media_t on all event listeners with MediaRef
The MediaRef passed via the listener is *temporary* and the client
application may not use it outside of the scope of the listener -
the Java object wouldn't be garbage collected, but the native
resource could become invalidated).

If a client application wants to keep the MediaRef, it should
invoke a factory method on the MediaRef to create a new instance.